Move some models to core
