+++ pom.xml
@@ -10,7 +10,7 @@
   </description>
   <url>https://github.com/Hygieia/${repository.name}</url>
   <packaging>jar</packaging>
-  <version>3.3.1-SNAPSHOT</version>
+  <version>3.3.2-SNAPSHOT</version>
 
   <parent>
     <groupId>org.springframework.boot</groupId>
@@ -48,7 +48,7 @@
 
   <properties>
     <!-- Dependencies -->
-    <com.capitalone.dashboard.core.version>3.14.0</com.capitalone.dashboard.core.version>
+    <com.capitalone.dashboard.core.version>3.14.1</com.capitalone.dashboard.core.version>
     <spring-security.version>4.2.18.RELEASE</spring-security.version>
     <tomcat.version>8.5.57</tomcat.version>
     <commons-beanutils.version>1.9.4</commons-beanutils.version>
