+++ src/main/java/com/capitalone/dashboard/collector/GitHubCollectorTask.java
@@ -12,8 +12,8 @@ import com.capitalone.dashboard.model.CollectorType;
 import com.capitalone.dashboard.model.Commit;
 import com.capitalone.dashboard.model.GitHubCollector;
 import com.capitalone.dashboard.model.GitHubParsed;
-import com.capitalone.dashboard.model.GitHubRepo;
 import com.capitalone.dashboard.model.GitRequest;
+import com.capitalone.dashboard.model.webhook.github.GitHubRepo;
 import com.capitalone.dashboard.repository.BaseCollectorRepository;
 import com.capitalone.dashboard.repository.CollectorItemMetadataRepository;
 import com.capitalone.dashboard.repository.CommitRepository;
